requirements.txt
aiohttp>=3.8.0
aiomysql>=0.1.0
python-dotenv>=0.21.0
typing-extensions>=4.0.0
